requirements.txt
flask
requests
sentence-transformers
faiss-cpu
